+++ xrlint/cli/main.py
@@ -2,7 +2,6 @@ import sys
 
 import click
 
-from xrlint.cli.stats import Stats
 
 # Warning: do not import heavy stuff here, it can
 # slow down commands like "xrlint --help" otherwise.
@@ -125,46 +124,41 @@ def main(
     Other inbuilt formats are `json` and `html` which you can specify
     using the `--format` option.
     """
-    from xrlint.cli.engine import CliEngine
+    from xrlint.cli.engine import XRLint
 
     if init_mode:
-        CliEngine.init_config_file()
-        raise click.exceptions.Exit(0)
+        XRLint.init_config_file()
+        return
 
-    cli_engine = CliEngine(
+    cli_engine = XRLint(
         no_config_lookup=no_config_lookup,
         config_path=config_path,
         plugin_specs=plugin_specs,
         rule_specs=rule_specs,
         output_format=output_format,
         output_path=output_file,
-        styled=color_enabled,
-        files=files,
+        output_styled=color_enabled,
+        max_warnings=max_warnings,
     )
-    stats = Stats()
 
-    config_list = cli_engine.load_config()
     if inspect_path:
-        import json
-
-        config = config_list.compute_config(inspect_path)
-        print(json.dumps(config.to_dict(), indent=2))
+        cli_engine.load_config_list()
+        cli_engine.print_config_for_file(inspect_path)
         return
 
-    if not files:
-        raise click.ClickException("No dataset files provided.")
+    if files:
+        cli_engine.load_config_list()
+        results = cli_engine.verify_datasets(files)
+        report = cli_engine.format_results(results)
+        cli_engine.write_report(report)
 
-    results = cli_engine.verify_datasets(config_list)
-    results = stats.collect(results)
-    report = cli_engine.format_results(results)
-    cli_engine.write_report(report)
-
-    error_status = stats.error_count > 0
-    max_warn_status = stats.warning_count > max_warnings
-    if max_warn_status and not error_status:
-        click.echo("maximum number of warnings exceeded.")
-    if max_warn_status or error_status:
-        raise click.exceptions.Exit(1)
+        result_stats = cli_engine.result_stats
+        error_status = result_stats.error_count > 0
+        max_warn_status = result_stats.warning_count > max_warnings
+        if max_warn_status and not error_status:
+            click.echo("Maximum number of warnings exceeded.")
+        if max_warn_status or error_status:
+            raise click.exceptions.Exit(1)
 
 
 if __name__ == "__main__":  # pragma: no cover